:test_tube: Refactored anoncreds routes tests
